Conflicts:
	src/compositor/texturing_gl.c
	src/compositor/visual_manager_3d_gl.c

+++ include/gpac/modules/service.h
@@ -253,7 +253,7 @@ typedef struct
 	u32 command_type;
 	LPNETCHANNEL on_channel;
 	//time in sec in the timeshift buffer - 0 means live point
-	u32 time;
+	Double time;
 } GF_NetComTimeShift;
